+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - [FEATURE] Add an experimental Core Animation recording pipeline for Session Replay, available through the `compositionTreeRecording` feature flag. See [#3127][]
+- [FEATURE] Add `disallowList` to `RUM.Configuration.URLSessionTracking` to exclude URLs from automatic RUM resource tracking, with `*` wildcard support. [#3097][]
 - [IMPROVEMENT] Forward `local_cache_hit` signal on RUM resources [#3074][]
 
 # 3.15.0 / 05-08-2026
@@ -1223,6 +1224,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#3109]: https://github.com/DataDog/dd-sdk-ios/pull/3109
 [#3074]: https://github.com/DataDog/dd-sdk-ios/pull/3074
 [#3127]: https://github.com/DataDog/dd-sdk-ios/pull/3127
+[#3097]: https://github.com/DataDog/dd-sdk-ios/pull/3097
 
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin